requirements.txt
torch
torchvision
pandas
scikit-learn
Pillow
fastapi
uvicorn
streamlit
kaggle
mlflow      # optional, for experiment logging
dvc         # optional, for pipeline versioning
python-multipart  # for file uploads in FastAPI
numpy<2.0
seaborn